refactor network configuration feature
